refactor(pages): replace about page with home and home page with profile
