+++ src/sections/profile/top-artists.spec.tsx
